refactor(plugin-navtree): extract shared getListActions helper
Address CodeRabbit review feedback:
- Extract the duplicated flatten/filter action-normalization pipeline into a
  shared `getListActions` helper in useActions.ts, used by both
  NavTreeItemColumns (rows) and L1Panel (header) so they can't drift.
- Import NavTreeItemColumns directly from its defining module in L1Panel
  instead of through the barrel (avoid re-exports per coding guidelines).
