+++ packages/plugins/plugin-navtree/src/hooks/useActions.ts
@@ -9,6 +9,16 @@ import { Graph, Node, useActions as useGraphActions } from '@dxos/plugin-graph';
 
 import { type FlattenedActions } from '#types';
 
+/**
+ * Flattens grouped actions and filters to the list-item dispositions rendered as
+ * navtree item actions. Shared by row (`NavTreeItemColumns`) and header (`L1Panel`)
+ * so both apply the same normalization rule.
+ */
+export const getListActions = ({ actions, groupedActions }: FlattenedActions): Node.Action[] =>
+  actions
+    .flatMap((action) => (Node.isAction(action) ? [action] : (groupedActions[action.id] ?? [])))
+    .filter((action) => ['list-item', 'list-item-primary'].includes(action.properties?.disposition));
+
 /** Returns flattened actions and grouped sub-actions for a given graph node. */
 export const useActions = (node: Node.Node): FlattenedActions => {
   const { graph } = useAppGraph();
